Keep a single class per a file
